Move devtools into subdirectory
